Moved quiesce impl to cpp
